Split events methods
